+++ Firmware/MotorControl/axis.cpp
@@ -11,18 +11,21 @@ Axis::Axis(const AxisHardwareConfig_t& hw_config,
            Encoder& encoder,
            SensorlessEstimator& sensorless_estimator,
            Controller& controller,
-           Motor& motor)
+           Motor& motor,
+           TrapezoidalTrajectory& trap)
     : hw_config_(hw_config),
       config_(config),
       encoder_(encoder),
       sensorless_estimator_(sensorless_estimator),
       controller_(controller),
-      motor_(motor)
+      motor_(motor),
+      trap_(trap)
 {
     encoder_.axis_ = this;
     sensorless_estimator_.axis_ = this;
     controller_.axis_ = this;
     motor_.axis_ = this;
+    trap_.axis_ = this;
 }
 
 static void step_cb_wrapper(void* ctx) {